Pass additional parameters to certbot
The Dockerfile calls the process with `ENTRYPOINT`: this allows the
caller to define arbitrary arguments that will be passed to the script.

The `run.sh` script is now calling certbot with `"$@"` in all cases:
this means that the additional arguments will be hopefully passed down
to certbot.

```
 # with-cotillons will be passed to certbot itself
docker run pierreprinetti/certbot with-cotillons
```

Closes #6